Yocto UDHCPC systemd integration part
For now we use the system and not embed a DHCPC client
library. This is aligned with future other configs that
could possibly be offered and involve applicative layers
protocols configuration (e.g. BGP to mention one).

One one hand, we offload the dhcp stateful functionality
externally (we can die hard, but the udhcpc dameon will
keep going).

On the other hand we tie ourself to yocto. In any case
those files will be placed in the yocyo recipes, so this
is quite legitimate.

One point on this: this stuff presumes we run the netconfd
in a non containerized fashion, since we need to access
the system infrastructure/facilities (busybox udhcpc).

We can rework this choice in the future if we foresee need for
frequent netconf updates (which are therefore not supported
in my opinion for now: you roll out a new machine for that).

In any case we have a whole bunch of things possibly involving the system OS for which we cannot operate in a straight
containerized environment unless with exceptions such as
volumes mount (e.g. we at least have to write /etc/resolv.conf).

We can try to go with containers for the sake of software updates maybe in the future, but we cannot fool ourselves into believing we are system/OS independent.
.